moved context to comp dir
